[staccmr-39] moving test to collections
